Fix #15

+++ scripts/build_n_push.sh
@@ -8,7 +8,7 @@
   echo "Branche : ${TRAVIS_BRANCH}" ;
   echo "PR : ${TRAVIS_PULL_REQUEST}" ;
   git commit -m "Deploy ${TRAVIS_BRANCH} to GitHub Pages" ;
-  if [ ${TRAVIS_BRANCH} == "master" ] 
+  if [ ${TRAVIS_BRANCH} == "master" ] && [ ${TRAVIS_PULL_REQUEST} == "false" ]
     then 
       echo "Push to julien-noblet/cad-killer";
       git push --force --quiet "https://${user}:${password}@github.com/julien-noblet/cad-killer.git" master:gh-pages;
